refactor(client): fold single-class index subpackages and move column stats index utils
- HoodieInMemoryHashIndex (index/inmemory), HoodieSimpleIndex and
  HoodieGlobalSimpleIndex (index/simple) move to org.apache.hudi.index root;
  the impls are instantiated via IndexType enum switch, not by FQN, and 1-2
  class subpackages added no grouping value. bloom/ and bucket/ stay.
- HoodieColumnStatsIndexUtils moves from org.apache.hudi.client to
  org.apache.hudi.metadata: it maintains the col-stats metadata index
  partition, beside HoodieTableMetadataUtil it delegates to.

Behavior-preserving; no moved class is @PublicAPIClass, config-FQN-referenced,
or Kryo-registered.
